[SCB-2610] refactor OperationException
