+++ clients/config-kie-client/src/main/java/org/apache/servicecomb/config/kie/client/KieConfigOperation.java
@@ -17,7 +17,7 @@
 
 package org.apache.servicecomb.config.kie.client;
 
-import org.apache.servicecomb.config.kie.client.exception.OperationException;
+import org.apache.servicecomb.config.common.exception.OperationException;
 import org.apache.servicecomb.config.kie.client.model.ConfigurationsRequest;
 import org.apache.servicecomb.config.kie.client.model.ConfigurationsResponse;
